# Improved Bake performance, 100x faster bake times, by eliminating unnecessary objects from the bake process.