# Jiggle Physics Baking Performance Optimization

| Property | Value |
|----------|-------|
| Date: | 2025-08-06 |
| Version: | Tested with Blender 4.5.0 LTS |

I'm attempting to diagnose a problem in this blender extension/addon.

The jiggle physics work for the most part. My only major problem so far is the baking logic. In my production Blend file, trying to bake the jiggle bones causes CPU to throttle at 100% for over an hour. The baking process is incredibly intensive and takes a long long time.

I have a theory on why this was happening, so I tested my hypothesis. My theory is that for some reason, the Jiggle Physics addon is recalculating the entire scene (all objects, all their modifiers, geonodes, everything) when it is baking each frame. This is incredibly unnecessary, as the only objects that should need to be calculated each frame are the armature itself and any collections/objects designated as collision in the addon panel.

My main production animation scene is quite complex with lots of modifiers and custom geometry node setups. So to test my theory I started a new blank blender file, and appended just the relevant armature and its actions. No meshes, no cameras, nothing else. When I use the Jiggle Physics addon to bake this scene, it bakes nearly instantly. So my theory is likely correct that the addon is recalculating EVERYTHING in a scene, which is overkill.

My final results are revealed at the bottom, but overall I'm seeing between a 100x and 300x performance boost, depending on the complexity of the scene.


## Problem Identification
- **Initial Issue**: Blender Jiggle Physics addon baking was extremely slow in production scenes with lots of complex modifiers, geometry, and nodes.
- **Symptoms**: 
  - Simple Test scene (2 armatures only): Fast baking, only a few seconds to bake.
  - Full Production scene (using those same 2 armatures + complex geometry/modifiers): 30 to 60 minutes to bake a short sequence.
  - Theory: I think the addon is recalculating the entire scene every frame during baking (including ALL geometry nodes, modifiers, meshes, etc).


## Investigation Process
### Step 1: Understanding the Root Cause
- Identified that `bpy.ops.nla.bake(visual_keying=True)` was the bottleneck.
- `visual_keying=True` forces Blender to evaluate the entire dependency graph for every frame.
- This includes all modifiers, geometry nodes, materials, and scene objects.
- The **Jiggle Physics** addon only needs armatures and collision objects, not the entire scene. Let's see if we can remove unnecessary objects from scene evaluation.

### Step 2: Apply the Correct Approach - Scene Context Optimization
- Optimize the NLA baking system, by limiting what gets evaluated during the NLA baking process.
- **Strategy**: Temporarily hide/disable expensive scene elements while keeping essential jiggle objects functional.
  - **Result**: I saw minor bake time improvements, but not major. This means *something* about my approach was working, but not fully.
  - **Next Step**: Implement some sort of NLA Baker debugger log so I can see where the bottleneck is.

### Step 3: Debugging and Analysis
- I implemented a comprehensive NLA Bake logging system to track exactly what was being evaluated, and how long each frame and each step was taking.
- **Critical discovery**:
  - There was a strange bottleneck right before the Bake process started. Some frames were being simulated, but very very slow (11+ seconds each).
  - Full Production scene: Only 3% of the NLA operation time was spent in `bpy.ops.nla.bake()`, 97% of effort was in the preroll phase.
- **Root cause identified**: Optimizations were applied too late. I had applied scene optimization to the bake phase, but the preroll phase was still evaluating the full scene, causing slow down.


## Final Solution
### Optimization Strategy
1. **Early scene optimization** - Apply aggressive optimization BEFORE preroll simulation, to improve baking speed.
2. **Aggressive object hiding** - Hide all non-essential objects:
   - *Keep*: Armatures, empties, designated collision objects.
   - *Hide*: All meshes, curves, lights, cameras (except collision objects).
3. **Modifier disabling** - Disable expensive modifiers even on remaining visible objects:
   - *Disabled*: Geometry Nodes, Subdivision, Multiresolution, Fluid, Cloth, etc. -- *everything non-essential!*

## Results
### Performance Improvements
I did not have a proper debugger implemented before my initial optimization, so the improvements recorded here are actually on the lower estimate side. But these are the results I got when implementing the initial bake optimizations, and then the final result when I moved the optimization steps to before the preroll phase.

#### **Test Production scene (48 frames)**:
Just a small 2 second sample from my main production scene.

| Test Phase | Results |
|------------|---------|
| Before Any Optimization: | (Several minutes bake time) |
| Before Preroll Optimization: | 73.549s bake time |
| After Preroll Optimization: | 3.247s bake time |
| **Improvement:** | **22.65x faster** at minimum |

#### **Real Production scene (456 frames)**:
The full timeline of my production scene.

| Test Phase | Results |
|------------|---------|
| Before Any Optimization: | (Several hours bake time) |
| Before Preroll Optimization: | 2150.506s (35+ minutes) bake time |
| Afte Preroll Optimization: | 19.965s bake time |
| **Improvement:** | **113x faster** at minimum |

### Technical Insights
- Jiggle physics modifies bone matrices directly via `VirtualParticle.apply_pose()`.
- `visual_keying=True` is required to capture these modifications in keyframes.
- Blender's dependency graph evaluation can be selectively optimized without breaking functionality of the Jiggle addon.
- Bake and Preroll simulation was consuming most of the time in complex scenes due to full scene evaluation.

## Final Process
The optimized baking process now:
1. Analyzes scene to identify essential objects (armatures + collision objects)
2. **Early optimization**: Hides non-essential objects and disables expensive modifiers
3. Runs preroll simulation with optimized scene
4. Executes `bpy.ops.nla.bake(visual_keying=True)` with optimized scene
5. Restores original scene state

This maintains full jiggle physics functionality while achieving 100x+ performance improvements in complex production scenes.

---

# Notes:
- I did not test this with bone collisions, as my scene did not require them. But there should be no issues in the logic, as the collision objects are specifically gathered and included in the Bake process.
- Further testing with collisions will be needed in the near future.
- The core baking logic wasn't really modified, so there should be no conflicts.
- Currently, this change simply gathers ALL armatures with jiggle physics enabled, and ALL collections/objects designated as collision. It could in theory be optimized further to only gather currently selected armatures, but I think this is so far a good first step.
- Overall, these improvements should be incorporated as soon as possible, as they save precious compute time and energy. Especially since the bake process seems to heavily rely on the slower CPU rather than GPU to do its work.