@naelstrof 2025-12-09
This looks well implemented and a great add, I'll look into merging it shortly. Sorry for the delay, I maintain lots of software.