@naelstrof 2025-12-09
Tested on Blender 5.0.0. Seems to work great. Good job saving the full state and restoring it properly, I was scared of doing something like that.

Exceptionally crafted pull request, thank you very much!